Move account and projects to own file as router
Also remove docs endpoint, since files don't exist
that it tries to render
