+++ lib/app.js
@@ -132,7 +132,6 @@ var init = exports.init = function (config) {
 
 
   // Routes
-
   app.get('/', routes.index);
   app.get('/about',function(req, res) {
      res.render('about');
@@ -145,11 +144,6 @@ var init = exports.init = function (config) {
   app.get('/styles/account-plugins-compiled.css', api_config.server('account', 'css'))
   app.get('/scripts/plugin-status-compiled.js', api_config.server('status', 'js'))
   app.get('/styles/plugin-status-compiled.css', api_config.server('status', 'css'))
-
-  // Docs page (single page for moment)
-  app.get('/docs', function(req, res) {res.render('docs');});
-  app.get('/docs/getting_started', function(req, res) {res.render('docs/GettingStarted');});
-  app.get('/docs/why_heroku_api_key', function(req, res) {res.render('docs/whyHerokuAPIKey');});
 
   app.get('/admin/projects', auth.requireAdminOr401, routes_admin.projects);
   app.get('/admin/users', auth.requireAdminOr401, routes_admin.users);
@@ -165,9 +159,8 @@ var init = exports.init = function (config) {
   app.get('/admin/plugins', auth.requireAdminOr401, routes_admin.plugins.get);
   app.put('/admin/plugins', auth.requireAdminOr401, routes_admin.plugins.put);
 
-  app.get('/account', auth.requireUser, routes.account);
-  app.get('/projects', auth.requireUser, routes.get_projects);
-  app.post('/projects', auth.requireUser, routes.post_projects);
+  app.use('/account', auth.requireUser, require('./routes/account'));
+  app.use('/projects', auth.requireUser, require('./routes/projects'));
 
   /* Requires at least read-only access to the repository in the path */
   app.get('/:org/:repo/', middleware.project, routes_jobs.html);
